refactor: move the ARIMA pipeline out of the application package
ADR 0004 established that fitting runs only in the offline precompute
batch, never on the request path, but the fitting code still lived in
`app/services/forecast.py` and imported statsmodels and scipy at module
scope. Those imports were the only reason either library was a runtime
dependency, so the property was behavioural rather than structural: the
request path happened not to fit, but nothing stopped it.

The pipeline moves to `backend/scripts/forecast/arima.py`, beside the
batch that is its only caller and outside the `app/` tree the Dockerfile
copies. `app/services/forecast.py` keeps only what the request path uses
— `is_eligible` (the ADR 0001 rule) and `build_response` — and imports
nothing that fits. `MIN_HISTORY_YEARS` stays with the eligibility rule
and is imported by the batch, so the threshold has one definition.

statsmodels and scipy move to the dev dependency group, which
`uv sync --frozen --no-dev` does not install, so they leave the runtime
image along with the pandas and patsy they pulled in.

Behaviour is unchanged. Forecasts and calibration built from the sample
database are byte-identical before and after the move (sha256
3d38e1e2…, 9 rows), and `make precompute-forecasts`, `make verify-db`
and `make publish-db` keep their interfaces, `--workers`/`--timeout`/
`--resume` included.

Tests:
- `test_runtime_dependencies.py` (new) imports `app.main` in a
  subprocess with statsmodels and scipy blocked from the import system,
  the way a container that never installed them behaves, and asserts
  neither is in `[project.dependencies]`.
- The two ARIMA-internals unit tests relocate with the code they cover,
  assertions unchanged, plus one covering the five internals the
  research harness's `current` arm composes, so the harness cannot break
  silently.
- `test_api.py`'s no-fitting-at-request-time guard keeps its assertions
  and re-points its monkeypatch at the batch module.
- `research/forecasting/methods.py` imports the pipeline at its new
  location rather than a copy, so rounds 1-6 stay reproducible.

Refs #41
