+++ backend/pyproject.toml
@@ -13,20 +13,25 @@ dependencies = [
     "pebble>=5.0.0",
     "pydantic>=2.0.0",
     "python-dotenv>=1.0.0",
-    "scipy>=1.11.0",
     "sentry-sdk[fastapi]>=2.68.1",
     "slowapi>=0.1.9",
-    "statsmodels>=0.14.0",
     "uvicorn[standard]>=0.52.4",
 ]
 
 [dependency-groups]
+# `uv sync --frozen --no-dev` builds the production image, so anything here is
+# absent from the container. The ARIMA fitting libraries belong here and not in
+# [project.dependencies]: fitting runs only in the offline batch
+# (scripts/forecast/), which the Dockerfile never copies. See
+# docs/adr/0004-forecasts-as-a-build-artifact.md.
 dev = [
     "httpx>=0.27.0",
     "pandas>=2.0.0",
     "pytest>=8.0.0",
     "ruff>=0.9.0",
+    "scipy>=1.11.0",
     "selenium>=4.0.0",
+    "statsmodels>=0.14.0",
 ]
 
 [tool.pytest.ini_options]
