+++ backend/tests/test_api.py
@@ -191,13 +191,13 @@ def test_forecast_endpoint_fits_no_model_at_request_time(monkeypatch):
     # Forecasts are precomputed by scripts/precompute_forecasts.py and stored;
     # the endpoint is a lookup. If it fit a model live, this would raise and
     # the request would 500 instead of returning a populated forecast.
-    from app.services import forecast
+    from scripts.forecast import arima
 
     def _must_not_be_called(*args, **kwargs):
         raise AssertionError("ARIMA fitting must not run on the request path")
 
-    monkeypatch.setattr(forecast, "_fit_best_model", _must_not_be_called)
-    monkeypatch.setattr(forecast, "fit_forecast", _must_not_be_called)
+    monkeypatch.setattr(arima, "_fit_best_model", _must_not_be_called)
+    monkeypatch.setattr(arima, "fit_forecast", _must_not_be_called)
 
     response = client.get("/api/names/emma/forecast", params={"sex": "F"})
     assert response.status_code == 200
